% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{activationFunIndex}
\alias{activationFunIndex}
\title{Assign ID to activation functions}
\usage{
activationFunIndex(funName)
}
\arguments{
\item{funName}{Type of activation function: "tanh", "sigm", "cdf", "relu" or
"softplus"}
}
\value{
An ID number which corresponds to:

\itemize{
\item 1 if "tanh"
}

\itemize{
\item 2 if "sigm"
}

\itemize{
\item 3 if "cdf"
}

\itemize{
\item 4 if "relu"
}

\itemize{
\item 5 if "softplus"
}
}
\description{
This function assigns a number attached to the type of activation function.
}
