% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{backwardParameterUpdate}
\alias{backwardParameterUpdate}
\title{Backward parameters update}
\usage{
backwardParameterUpdate(mp, Sp, mzF, SzF, SzB, Czp, mzB, idx)
}
\arguments{
\item{mp}{Mean vector of the parameters for the current layer}

\item{Sp}{Covariance matrix of the parameters for the current layer}

\item{mzF}{Mean vector of the units for the next layer}

\item{SzF}{Covariance matrix of the units for the next layer}

\item{SzB}{Updated covariance matrix of the units for the next layer}

\item{Czp}{Covariance matrix between units and parameters for the current layer}

\item{mzB}{Updated mean vector of the units for the next layer}

\item{idx}{List that contains the indices for the parameter update step of
the current layer}
}
\value{
A list that contains:

\itemize{
\item Updated mean vector of the parameters for the current layer
}

\itemize{
\item Updated covariance matrix of the parameters for the current layer
}
}
\description{
This function updates parameters from responses to input data. It updates
\eqn{\mu_{\theta|y}} and \eqn{\Sigma_{\theta|y}} from the \eqn{\theta|y}
distribution for a given layer.
}
