% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{batchDerivative}
\alias{batchDerivative}
\title{One iteration of the TAGI with derivative calculations}
\usage{
batchDerivative(NN, theta, normStat, states, x, Sx, y, dlayer)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{theta}{List of parameters}

\item{normStat}{Normalized statistics}

\item{states}{List of states}

\item{x}{Input data}

\item{Sx}{Variance of input data}

\item{y}{Response data}

\item{dlayer}{Layer from which derivatives will be in respect to}
}
\value{
\itemize{
\item List of parameters
}

\itemize{
\item List of normalized statistics
}

\itemize{
\item Mean of predicted responses
}

\itemize{
\item Variance of predicted responses
}

\itemize{
\item Mean of first derivative of predicted responses
}

\itemize{
\item Variance of first derivative of predicted responses
}

\itemize{
\item Covariance between derivatives and inputs
}

\itemize{
\item Mean of second derivative of predicted responses
}
}
\description{
This function goes through one learning iteration of the neural network model
using TAGI with derivative calculations.
}
