% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BH}
\alias{BH}
\title{Price of 506 Boston houses.}
\format{
A data frame with 506 rows and 14 variables:
\describe{
\item{CRIM }{per capita crime rate by town}
\item{ZN}{proportion of residential land zoned for lots over 25,000 sq.ft.}
\item{INDUS}{proportion of non-retail business acres per town}
\item{CHAS}{Charles River dummy variable (= 1 if tract bounds river; 0 otherwise)}
\item{NOX}{nitric oxides concentration (parts per 10 million)}
\item{RM}{average number of rooms per dwelling}
\item{AGE}{proportion of owner-occupied units built prior to 1940}
\item{DIS}{weighted distances to five Boston employment centres}
\item{RAD}{index of accessibility to radial highways}
\item{TAX}{full-value property-tax rate per $10,000}
\item{PTRATIO}{pupil-teacher ratio by town}
\item{B}{1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town}
\item{LSTAT}{\% lower status of the population}
\item{MEDV}{median value of owner-occupied homes in $1000's}
}
}
\source{
\url{https://github.com/CivML-PolyMtl/TAGI/blob/master/BostonHousing/data/BostonHousing.mat}
}
\usage{
BH
}
\description{
This dataset was originally from the StatLib archive. It contains the price
and other attributes of 506 Boston houses.
}
\details{
The dataset from the TAGI repository was used for comparison purposes,
but the original dataset was published by Harrison, D. and Rubinfeld, D.L.
}
\references{
http://lib.stat.cmu.edu/datasets/boston

Harrison, D. and Rubinfeld, D.L. `Hedonic prices and the demand
for clean air', J. Environ. Economics & Management, vol.5, 81-102, 1978.
}
\keyword{datasets}
