% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{buildCzp}
\alias{buildCzp}
\title{Reformat covariance matrix between units and parameters}
\usage{
buildCzp(Czw, Czb, currentHiddenUnit, prevHiddenUnit, batchSize)
}
\arguments{
\item{Czw}{Covariance matrix between units and weights for the current layer}

\item{Czb}{Covariance matrix between units and baises for the current layer}

\item{currentHiddenUnit}{Number of units in the current layer}

\item{prevHiddenUnit}{Number of units in the previous layer}

\item{batchSize}{Number of observations trained at the same time}
}
\value{
Reformatted covariance matrix between units and parameters
}
\description{
This function properly reformats covariance matrix between units and parameters
\eqn{\Sigma_{Z\theta}} for the update step.
}
