% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{buildCzz}
\alias{buildCzz}
\title{Reformat covariance matrix between units of the previous and current layers}
\usage{
buildCzz(Czz, currentHiddenUnit, prevHiddenUnit, batchSize)
}
\arguments{
\item{Czz}{Covariance matrix between units of the previous and current layers}

\item{currentHiddenUnit}{Number of units in the current layer}

\item{prevHiddenUnit}{Number of units in the previous layer}

\item{batchSize}{Number of observations trained at the same time}
}
\value{
Reformatted covariance matrix between of the previous and current layers
}
\description{
This function properly reformats covariance matrix between units of the
previous and current layers \eqn{\Sigma_{ZZ^{+}}} for the update step.
}
