% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{catParameters}
\alias{catParameters}
\title{Concatenate parameters}
\usage{
catParameters(mw, Sw, mb, Sb, mwx, Swx, mbx, Sbx)
}
\arguments{
\item{mw}{Mean of weights for the current layer}

\item{Sw}{Covariance of weights for the current layer}

\item{mb}{Mean of biases for the current layer}

\item{Sb}{Covariance of biases for the current layer}

\item{...}{Other parameters}
}
\value{
\itemize{
\item Mean vector of weights for the current layer
}

\itemize{
\item Covariance vector of weights for the current layer
}

\itemize{
\item Mean vector of biases for the current layer
}

\itemize{
\item Covariance vector of biases for the current layer
}
}
\description{
Combines in a single column vector each parameter for all layers.
}
