% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{compressNormStat}
\alias{compressNormStat}
\title{Compress Normalized Statistics TBD}
\usage{
compressNormStat(mra, Sra)
}
\arguments{
\item{mra}{TBD}

\item{Sra}{TBD}
}
\value{
normStat: TBD
}
\description{
Put together normalized statistics into a list.
}
