% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{compressParameters}
\alias{compressParameters}
\title{Compress parameters}
\usage{
compressParameters(mw, Sw, mb, Sb, mwx, Swx, mbx, Sbx)
}
\arguments{
\item{mw}{Mean vector of weights for the current layer}

\item{Sw}{Covariance vector of weights for the current layer}

\item{mb}{Mean vector of biases for the current layer}

\item{Sb}{Covariance vector of biases for the current layer}

\item{...}{Other parameters}
}
\value{
List of parameters
}
\description{
Put together parameters into a list of parameters.
}
