% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{compressStates}
\alias{compressStates}
\title{Compress states}
\usage{
compressStates(mz, Sz, ma, Sa, J, mdxs, Sdxs, mxs, Sxs)
}
\arguments{
\item{mz}{Mean of units for each layer}

\item{Sz}{Covariance of units for each layer}

\item{ma}{Mean of activated units for each layer}

\item{Sa}{Covariance of activated units for each layer}

\item{J}{Jacobian}

\item{...}{Other parameters}
}
\value{
List of states
}
\description{
Put together states into a list of states.
}
