% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{covariance}
\alias{covariance}
\title{Indices for covariances in the neural network}
\usage{
covariance(NN)
}
\arguments{
\item{NN}{Lists the structure of the neural network}
}
\value{
NN with new elements:

\itemize{
\item Indices (weights and activation units) for
deterministic matrix F *  \eqn{\mu_{WA}} for each layer
}

\itemize{
\item Bias indices for deterministic matrix F *  \eqn{\mu_{B}} for
each layer
}

\itemize{
\item Indices (weights and activation units) for
deterministic matrix F *  \eqn{\Sigma_{ZWA}} for each layer
}

\itemize{
\item Indices for the parameter update step for each layer
}

\itemize{
\item Indices for the hidden state update step for each
layer
}

\itemize{
\item Indices (weights and activation units) for deterministic matrix F *  \eqn{\Sigma_{WA\theta}} for
each layer
}

\itemize{
\item Indices for activation unit for each layer
}

\itemize{
\item Bias indices for deterministic matrix F *  \eqn{\Sigma_{B}} for
each layer
}
}
\description{
This function assigns indices for all covariance elements in
the neural network.
}
