% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{covarianceCzz}
\alias{covarianceCzz}
\title{Covariance matrix between units of the previous and current layers}
\usage{
covarianceCzz(mp, Sz, J, idxCawa)
}
\arguments{
\item{mp}{Mean vector of the parameters for the current layer}

\item{Sz}{Covariance matrix of the units for the current layer}

\item{J}{Jacobian matrix evaluated at \eqn{\mu_{z}}}

\item{idxCawa}{List that contains the indices for weights and for activation
units for the current and previous layers respectively}
}
\value{
Covariance matrix between units of the previous and current layers
}
\description{
This function calculate the covariance matrix between units of the previous
and current layers \eqn{\Sigma_{ZZ^{+}}} for a given layer.
}
