% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{covarianceSz}
\alias{covarianceSz}
\title{Covariance matrix of units}
\usage{
covarianceSz(mp, ma, Sp, Sa, idxFSwaF, idxFSwaFb)
}
\arguments{
\item{mp}{Mean vector of the parameters for the current layer}

\item{ma}{Mean vector of the activation units from previous layer}

\item{Sp}{Covariance matrix of the parameters for the current layer}

\item{Sa}{Covariance matrix of the activation units from previous layer}

\item{idxFSwaF}{List that contains the indices for weights and for activation
units for the current and previous layers respectively}

\item{idxFSwaFb}{Indices for biases of the current layer}
}
\value{
Covariance matrix of units for the current layer
}
\description{
This function calculate the covariance matrix of the units \eqn{\Sigma_{Z}} for a given layer.
}
