% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{covdx}
\alias{covdx}
\title{Covariance between derivatives and hidden states}
\usage{
covdx(mwo, mw, mdgo2, mpdi, mdgoe, Cdozi, Cdizi, ni, no, no2, B)
}
\arguments{
\item{mwo}{Mean vector of weights for the next layer}

\item{mw}{Mean vector of weights for the current layer}

\item{mdgo2}{Mean vector of product of derivatives in second next layer}

\item{mpdi}{Mean vector of first derivative product wd of current layer}

\item{mdgoe}{Mean of product of derivatives at each node in next layer}

\item{Cdozi}{Covariance between derivative (next) and hidden (current) layers}

\item{Cdizi}{Covariance between derivative and hidden layers (same layer)}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}
}
\value{
Covariance between derivative and hidden states
}
\description{
This function calculates covariance between derivatives and hidden states.
It is not related to the derivative calculation process.
It could be used infer Z (hidden states) with the constraint that the derivative of g with respect to Z equals 0.
}
