% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{createInitCellwithArray}
\alias{createInitCellwithArray}
\title{Initialization (matrix of lists)}
\usage{
createInitCellwithArray(numLayers)
}
\arguments{
\item{numLayers}{Number of layers in the neural network}
}
\value{
Matrix containing empty lists
}
\description{
Initializes a matrix containing lists.
}
