% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{createInitNormStat}
\alias{createInitNormStat}
\title{Initialization (TBD)}
\usage{
createInitNormStat(NN)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}
}
\value{
normStat: TBD
}
\description{
Initializes a matrix containing lists of 0 for TBD.
}
