% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{createStateCellarray}
\alias{createStateCellarray}
\title{States initialization (zero-matrices)}
\usage{
createStateCellarray(nodes, numLayers, B, rB)
}
\arguments{
\item{nodes}{Vector which contains the number of nodes at each layer}

\item{numLayers}{Number of layers in the neural network}

\item{B}{Batch size}

\item{rB}{Number of times batch size is repeated}
}
\value{
Zero-matrices for each layer
}
\description{
Initiliazes neural network states at 0.
}
