% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{derivative}
\alias{derivative}
\title{Derivative calculation}
\usage{
derivative(NN, theta, states, mda, Sda, mdda, Sdda, dlayer)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{theta}{List of parameters}

\item{states}{List of states}

\item{mda}{Mean vectors of activation units' first derivative}

\item{Sda}{Covariance matrices of activation units' first derivative}

\item{mdda}{Mean vectors of activation units' second derivative}

\item{Sdda}{Covariance matrices of activation units' second derivative}

\item{dlayer}{layer from which derivatives will be in respect to}
}
\value{
\itemize{
\item Mean of first derivative of predicted responses
}

\itemize{
\item Variance of first derivative of predicted responses
}

\itemize{
\item Covariance between derivatives and inputs
}

\itemize{
\item Mean of second derivative of predicted responses
}
}
\description{
This function does derivative calculations.
}
