% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{extractParameters}
\alias{extractParameters}
\title{Extract Parameters}
\usage{
extractParameters(theta)
}
\arguments{
\item{theta}{List of parameters}
}
\value{
mw: TBD

Sw: TBD

mb: TBD

Sb: TBD

mwx: TBD

Swx: TBD

mbx: TBD

Sbx: TBD
}
\description{
Extract parameters from list of parameters.
}
