% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{extractStates}
\alias{extractStates}
\title{Extract states}
\usage{
extractStates(states)
}
\arguments{
\item{states}{List of states}
}
\value{
\itemize{
\item Mean of units for each layer
}

\itemize{
\item Covariance of units for each layer
}

\itemize{
\item Mean of activated units for each layer
}

\itemize{
\item Covariance of activated units for each layer
}

\itemize{
\item Jacobian
}
}
\description{
Extract states from list of states.
}
