% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovaddddddw}
\alias{fcCovaddddddw}
\title{Covariance between first and second derivatives from consecutive layers}
\usage{
fcCovaddddddw(mao, mai, mdao, Caoai, Cdodi, Caow, Cdow, acto, acti, ni, no, B)
}
\arguments{
\item{mao}{Mean vector of activation units from next layer}

\item{mai}{Mean vector of activation units from current layer}

\item{mdao}{Mean vector of activation units' first derivative from next layer}

\item{Caoai}{Covariance between activation units from current and next layers}

\item{Cdodi}{Covariance between first derivatives from current and next layers}

\item{Caow}{Covariance between activation units and weights}

\item{Cdow}{Covariance between derivatives and weights}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
\itemize{
\item Covariance between first and second derivatives from consecutive layers
}

\itemize{
\item Covariance between second derivatives from next layer and weights
}
}
\description{
This function calculates covariance between weights and second derivatives and
covariance between first and second derivatives from consecutive layers.
}
