% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovdaddd}
\alias{fcCovdaddd}
\title{Covariance between first and second derivatives from consecutive layers}
\usage{
fcCovdaddd(mao, mai, mdai, Caoai, Cdodi, acti, ni, no, B)
}
\arguments{
\item{mao}{Mean vector of activation units from next layer}

\item{mai}{Mean vector of activation units from current layer}

\item{mdai}{Mean vector of activation units' first derivative from current layer}

\item{Caoai}{Covariance between activation units from current and next layers}

\item{Cdodi}{Covariance between derivatives from current and next layers}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
\itemize{
\item Covariance between first derivatives from next layer and activation units from current layer
}

\itemize{
\item Covariance between first and second derivatives from consecutive layers
}
}
\description{
This function calculates covariance between activation units and first derivatives and
covariance between first and second derivatives from consecutive layers.
}
