% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovDlayer}
\alias{fcCovDlayer}
\title{Covariance between Products of Derivatives and Weights}
\usage{
fcCovDlayer(mdgo2, mwo, Cdowdi, ni, no, no2, B)
}
\arguments{
\item{mdgo2}{Mean vector of derivatives in 2nd next layer}

\item{mwo}{Mean vector of the weights for the next layer}

\item{Cdowdi}{Covariance between derivatives and weights times derivatives}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}
}
\value{
Covariance between weights times derivatives from consecutive layers
}
\description{
This function calculates covariance between products of derivatives and
weights from consecutive layers.
}
