% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovdwd}
\alias{fcCovdwd}
\title{Covariance between Derivatives and Weights*Derivatives}
\usage{
fcCovdwd(md, mw, Cdow, Cdodi, ni, no, B)
}
\arguments{
\item{md}{Mean vector of derivatives}

\item{mw}{Mean vector of the weights for the current layer}

\item{Cdow}{Covariance between derivatives and weights}

\item{Cdodi}{Covariance between derivatives from current and next layers}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Covariance between derivatives and weights times derivatives
}
\description{
This function calculates covariance between derivatives and weights and
covariance between derivatives from consecutive layers.
}
