% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovdwddd}
\alias{fcCovdwddd}
\title{Covariance between Derivatives and Weights}
\usage{
fcCovdwddd(mao, Sao, mai, Sai, Caow, Caoai, acto, acti, ni, no, B)
}
\arguments{
\item{mao}{Mean vector of the activation units from next layer}

\item{Sao}{Covariance of the activation units from next layer}

\item{mai}{Mean vector of the activation units from current layer}

\item{Sai}{Covariance of the activation units from current layer}

\item{Caow}{Covariance between activation units and weights}

\item{Caoai}{Covariance between activation units from current and next layers}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Covariance between derivatives and weights

Covariance between derivatives from current and next layers
}
\description{
This function calculates covariance between derivatives and weights and
covariance between derivatives from consecutive layers.
}
