% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCwdowdowwdi22}
\alias{fcCwdowdowwdi22}
\title{Covariance between Next Layer Multiplied Products and Current Layer Multiplied Products (Same Derivative)}
\usage{
fcCwdowdowwdi22(mpdi, mpdo, mdgo2, Cdgodgi, ni, no, no2, B)
}
\arguments{
\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{mpdo}{Mean vector of the first order derivative product wd of next layer}

\item{mdgo2}{Mean vector of derivatives in 2nd next layer}

\item{Cdgodgi}{Covariance between weights times derivatives from consecutive layers}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}
}
\value{
Covariance cov(wdowdo,wdiwdi) where the di terms are the same
}
\description{
This function calculates covariance cov(wdowdo,wdiwdi) where the di terms are the same when next second layer involves multiplied terms (wdo2wdo2).
}
