% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCwdowdowwdi2_3hl}
\alias{fcCwdowdowwdi2_3hl}
\title{Covariance between next layer multiplied products and current layer multiplied products (same derivative, minimum 3 hidden layers)}
\usage{
fcCwdowdowwdi2_3hl(mpdi, mpdo, mdgo2, Cdgodgi, acti, ni, no, no2, B)
}
\arguments{
\item{mpdi}{Mean vector of first derivative product wd of current layer}

\item{mpdo}{Mean vector of first derivative product wd of next layer}

\item{mdgo2}{Mean vector of product of derivatives in second next layer}

\item{Cdgodgi}{Covariance between weights times derivatives from consecutive layers}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}
}
\value{
Covariance cov(wdowdo,wdiwdi) where the di terms are the same
}
\description{
This function calculates covariance cov(wdowdo,wdiwdi) where the di terms are the same when next second layer involves multiplied terms (wdo2wdo2).
It is used when there are at least 3 hidden layers.
}
