% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDerivative}
\alias{fcDerivative}
\title{Derivatives for fully connected layers}
\usage{
fcDerivative(
  mw,
  Sw,
  mwo,
  Jo,
  J,
  mao,
  Sao,
  mai,
  Sai,
  Szi,
  mdai,
  Sdai,
  mdgo,
  mdgoe,
  Sdgo,
  mdgo2,
  acto,
  acti,
  ni,
  no,
  no2,
  B
)
}
\arguments{
\item{mw}{Mean vector of weights for the current layer}

\item{Sw}{Covariance of weights for the current layer}

\item{mwo}{Mean vector of weights for the next layer}

\item{Jo}{Jacobian of next layer}

\item{J}{Jacobian of current layer}

\item{mao}{Mean vector of activation units from next layer}

\item{Sao}{Covariance of activation units from next layer}

\item{mai}{Mean vector of activation units from current layer}

\item{Sai}{Covariance of activation units from current layer}

\item{Szi}{Covariance of units from current layer}

\item{mdai}{Mean vector of activation units' first derivative from current layer}

\item{Sdai}{Covariance of activation units' first derivative from current layer}

\item{mdgo}{Mean vector of product of derivatives in next layer}

\item{mdgoe}{Mean of product of derivatives at each node in next layer}

\item{Sdgo}{Variance of first derivatives in next layer}

\item{mdgo2}{Mean vector of product of derivatives in second next layer}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}
}
\value{
Mean vector of first derivatives

Covariance matrix of first derivatives

Covariance matrix of first derivative and input layer

Covariance between activation units and weights

Covariance between activation units from current and next layers

Covariance between first derivatives and weights

Covariance between first derivatives from current and next layers

Covariance between first derivatives from next layer and weights times derivatives from current layer
}
\description{
This function calculates mean and variance of derivatives and covariance of derivative and input layers.
}
