% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDerivative2}
\alias{fcDerivative2}
\title{Second Derivatives for Fully Connected layers}
\usage{
fcDerivative2(
  mw,
  mwo,
  mao,
  mai,
  mdai,
  mddai,
  mpddi,
  mdgo,
  mdgo2,
  Caow,
  Caoai,
  Cdow,
  Cdodi,
  acti,
  ni,
  no,
  no2,
  B
)
}
\arguments{
\item{mw}{Mean vector of the weights for the current layer}

\item{mwo}{Mean vector of the weights for the next layer}

\item{mao}{Mean vector of the activation units from next layer}

\item{mai}{Mean vector of the activation units from current layer}

\item{mdai}{Mean vector of the activation units' derivative from current layer}

\item{mddai}{Mean vector of the activation units' second derivative from current layer}

\item{mpddi}{Mean vector of the second order derivative product wdd of current layer}

\item{mdgo}{Mean vector of derivatives in next layer}

\item{mdgo2}{Mean vector of derivatives in 2nd next layer}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}
}
\value{
Mean vector of the derivatives
}
\description{
This function calculates mean of product of derivatives, when new product term involves second order derivatives (wdd).
}
