% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDerivative3}
\alias{fcDerivative3}
\title{Products of First Order Derivatives Multiplied to Second Order Derivative for Fully Connected layers}
\usage{
fcDerivative3(
  mw,
  mwo,
  mao,
  mai,
  mdao,
  mdai,
  mpdi,
  mpdi2,
  mdgo,
  mdgo2,
  Caow,
  Caoai,
  Cdow,
  Cdodi,
  acto,
  acti,
  ni,
  no,
  no2,
  B
)
}
\arguments{
\item{mw}{Mean vector of the weights for the current layer}

\item{mwo}{Mean vector of the weights for the next layer}

\item{mao}{Mean vector of the activation units from next layer}

\item{mai}{Mean vector of the activation units from current layer}

\item{mdao}{Mean vector of the activation units' derivative from next layer}

\item{mdai}{Mean vector of the activation units' derivative from current layer}

\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{mpdi2}{Mean array of combination of products of first order derivatives}

\item{mdgo}{Mean vector of derivatives in next layer}

\item{mdgo2}{Mean vector of derivatives in 2nd next layer}

\item{Caow}{Covariance between activation units and weights}

\item{Caoai}{Covariance between activation units from current and next layers}

\item{Cdow}{Covariance between derivatives and weights}

\item{Cdodi}{Covariance between derivatives from current and next layers}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}
}
\value{
Mean vector of the derivatives
}
\description{
This function calculates mean of product of derivatives, when new product term involves product
of two first order derivatives from the same layer (wd*wd) multiplied to second order derivatives (wdd) from next layer.
}
