% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDerivative4}
\alias{fcDerivative4}
\title{Products of First Order Derivatives Multiplied to Products of First Order Derivatives for Fully Connected layers}
\usage{
fcDerivative4(
  mw,
  Sw,
  mwo,
  mao,
  mai,
  mdao,
  mdai,
  Sdai,
  mpdo,
  mpdi,
  mdgo,
  mdgo2,
  Cdgodgi,
  acto,
  acti,
  ni,
  no,
  no2,
  B,
  dlayer
)
}
\arguments{
\item{mw}{Mean vector of the weights for the current layer}

\item{Sw}{Covariance of the weights for the current layer}

\item{mwo}{Mean vector of the weights for the next layer}

\item{mao}{Mean vector of the activation units from next layer}

\item{mai}{Mean vector of the activation units from current layer}

\item{mdao}{Mean vector of the activation units' derivative from next layer}

\item{mdai}{Mean vector of the activation units' derivative from current layer}

\item{Sdai}{Covariance of the activation units' derivative from current layer}

\item{mpdo}{Mean vector of the first order derivative product wd of next layer}

\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{mdgo}{Mean vector of derivatives in next layer}

\item{mdgo2}{Mean vector of derivatives in 2nd next layer}

\item{Cdgodgi}{Covariance between weights times derivatives from consecutive layers}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}

\item{dlayer}{TRUE if layer from which derivatives will be in respect to}
}
\value{
Mean vector of the derivatives
}
\description{
This function calculates mean of product of derivatives, when new product term involves product
of two first order derivatives (wd\emph{wd) from the same layer multiplied to product
of two first order derivatives (wd}wd) from next layer.
}
