% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDlayer_actDerivative}
\alias{fcDlayer_actDerivative}
\title{Expectation and Covariance at Input Layer Using Derivative of Previous Order}
\usage{
fcDlayer_actDerivative(
  mpdi,
  Spdi,
  mdgo,
  mdgoe,
  Sdgoe,
  mw,
  Sw,
  Cdgodgi,
  ni,
  no,
  no2,
  B
)
}
\arguments{
\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{Spdi}{Covariance of the first order derivative product wd of current layer}

\item{mdgo}{Mean vector of derivatives in next layer}

\item{mdgoe}{Mean derivatives at each node in next layer}

\item{Sdgoe}{Covariance derivatives at each node in next layer}

\item{mw}{Mean vector of the weights for the next layer}

\item{Sw}{Covariance of the weights for the current layer}

\item{Cdgodgi}{Covariance between weights times derivatives from consecutive layers}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}
}
\value{
Covariance between weights times derivatives from consecutive layers
}
\description{
This function calculates the expectation of products of current layer and
covariance between products of derivatives and
weights from consecutive layers where the derivative from current layer is the derivative result of previous order.
}
