% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcHiddenStateBackwardPass}
\alias{fcHiddenStateBackwardPass}
\title{Backpropagation (states' deltas) for fully connected layers (many observations)}
\usage{
fcHiddenStateBackwardPass(Sz, Sxs, J, mw, deltaM, deltaS, ni, no, B, rB)
}
\arguments{
\item{Sz}{Covariance of units from current layer}

\item{Sxs}{Null by default (not used yet)}

\item{J}{Jacobian of current layer}

\item{mw}{Mean vector of weights for the current layer}

\item{deltaM}{Delta of mean vector of the next layer units given \eqn{y} \eqn{\mu_{Z}|y}}

\item{deltaS}{Delta of covariance matrix of the next layer units given \eqn{y} \eqn{\Sigma_{Z}|y}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}

\item{rB}{Number of times batch size is repeated}
}
\value{
\itemize{
\item Delta of mean vector of current layer units given \eqn{y} \eqn{\mu_{Z}|y}
}

\itemize{
\item Delta of covariance matrix of current layer units given \eqn{y} \eqn{\Sigma_{Z}|y}
}
}
\description{
This function calculates units' deltas at a given layer when using more than one
observation at the time.
}
