% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcHiddenStateBackwardPassB1}
\alias{fcHiddenStateBackwardPassB1}
\title{Backpropagation (states' deltas) for fully connected layers (one observation)}
\usage{
fcHiddenStateBackwardPassB1(Sz, Sxs, J, mw, deltaM, deltaS, ni, no)
}
\arguments{
\item{Sz}{Covariance of the units from current layer}

\item{Sxs}{Null by default (not used yet)}

\item{J}{Jacobian of current layer}

\item{mw}{Mean vector of weights for the current layer}

\item{deltaM}{Delta of mean vector of next layer units given \eqn{y} \eqn{\mu_{Z}|y}}

\item{deltaS}{Delta of covariance matrix of next layer units given \eqn{y} \eqn{\Sigma_{Z}|y}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}
}
\value{
\itemize{
\item Delta of mean vector of current layer units given \eqn{y} \eqn{\mu_{Z}|y}
}

\itemize{
\item Delta of covariance matrix of current layer units given \eqn{y} \eqn{\Sigma_{Z}|y}
}
}
\description{
This function calculates units' deltas at a given layer when using one observation at the time.
}
