% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcMeanDlayer2row}
\alias{fcMeanDlayer2row}
\title{Mean of Weights times Derivatives Products Terms Squared (wdo x (wdi*wdi))}
\usage{
fcMeanDlayer2row(mpdi, mpdi2, mdgo, Cwdowdiwdi, ni, no, no2, B)
}
\arguments{
\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{mpdi2}{Mean array of combination of products of first order derivatives}

\item{mdgo}{Mean vector of derivatives in next layer}

\item{Cwdowdiwdi}{Covariance cov(wdo,(wdi*wdi)) of weights times derivatives products terms when there is one product in next layer and two in current}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}
}
\value{
Mean of weights times derivatives products terms
}
\description{
This function calculates mean of weights times derivatives products terms when
adding two of those products from current layer to already calculated expectation
that ended with one such product of next layer (i.e. wdo x (wdi\emph{wdi)). Mean and
covariance terms are in array format. Once added, rows need to be
summed to aggregate expectations by node}node combinations of current layer.
}
