% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcMeanVarB1}
\alias{fcMeanVarB1}
\title{Mean and covariance vectors of units (one observation)}
\usage{
fcMeanVarB1(mw, Sw, mb, Sb, ma, Sa, ni, no)
}
\arguments{
\item{mw}{Mean vector of weights for the current layer}

\item{Sw}{Covariance of weights for the current layer}

\item{mb}{Mean vector of biases for the current layer}

\item{Sb}{Covariance of biases for the current layer}

\item{ma}{Mean vector of activation units from previous layer}

\item{Sa}{Covariance of activation units from previous layer}

\item{ni}{Number of units in previous layer}

\item{no}{Number of units in current layer}
}
\value{
\itemize{
\item Mean vector of units for the current layer \eqn{\mu_{Z}}
}

\itemize{
\item Covariance matrix of units for the current layer \eqn{\Sigma_{Z}}
}
}
\description{
This function calculate the mean vector of units \eqn{\mu_{Z}} and the covariance matrix of the units \eqn{\Sigma_{Z}}for a given layer.
}
