% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcMeanVarDlayer}
\alias{fcMeanVarDlayer}
\title{Mean and variance of weights times derivatives products terms}
\usage{
fcMeanVarDlayer(mx, Sx, my, mye, Sy, Cxy, ni, no, no2, B)
}
\arguments{
\item{mx}{Mean vector of inputs}

\item{Sx}{Variance of inputs}

\item{my}{Mean vector of outputs}

\item{mye}{Mean derivatives at each node in next layer}

\item{Sy}{Variance of outputs}

\item{Cxy}{Covariance between inputs and outputs}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}
}
\value{
\itemize{
\item Mean of weights times derivatives products terms
}

\itemize{
\item Covariance between weights times derivatives products terms
}
}
\description{
This function calculates mean and variance of weights times derivatives products terms.
}
