% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcParameterBackwardPassB1}
\alias{fcParameterBackwardPassB1}
\title{Backpropagation (parameters' deltas) for fully connected layers (one observation)}
\usage{
fcParameterBackwardPassB1(Sw, Sb, ma, deltaMr, deltaSr, ni, no)
}
\arguments{
\item{Sw}{Covariance of weights for the current layer}

\item{Sb}{Covariance of biaises for the current layer}

\item{ma}{Mean vector of activation units for the current layer}

\item{deltaMr}{Delta of mean vector of next layer units given \eqn{y} \eqn{\mu_{Z}|y}}

\item{deltaSr}{Delta of covariance matrix of next layer units given \eqn{y} \eqn{\Sigma_{Z}|y}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}
}
\value{
\itemize{
\item Delta of mean vector of weights given \eqn{y} \eqn{\mu_{\theta}|y}
}

\itemize{
\item Delta of covariance matrix of weights given \eqn{y} \eqn{\Sigma_{\theta}|y}
}

\itemize{
\item Delta of mean vector of biases given \eqn{y} \eqn{\mu_{\theta}|y}
}

\itemize{
\item Delta of covariance matrix of biaises given \eqn{y} \eqn{\Sigma_{\theta}|y}
}
}
\description{
This function calculates parameters' deltas at a given layer when using one observation at the time.
}
