% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{feedBackward}
\alias{feedBackward}
\title{Backpropagation}
\usage{
feedBackward(NN, mp, Sp, mz, Sz, Czw, Czb, Czz, y)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{mp}{Mean vectors of parameters for each layer \eqn{\mu_{\theta}}}

\item{Sp}{Covariance matrices of parameters for each layer \eqn{\Sigma_{\theta}}}

\item{mz}{Mean vectors of units for each layer \eqn{\mu_{Z}}}

\item{Sz}{Covariance matrices of units for each layer \eqn{\Sigma_{Z}}}

\item{Czw}{Covariance matrices between units and weights for each layer \eqn{\Sigma_{ZW}}}

\item{Czb}{Covariance matrices between units and biases for each layer \eqn{\Sigma_{ZB}}}

\item{Czz}{Covariance matrices between previous and current units for each layer \eqn{\Sigma_{ZZ^{+}}}}

\item{y}{Response data}
}
\value{
\itemize{
\item Updated mean vectors of parameters for each layer \eqn{\mu_{\theta}}
}

\itemize{
\item Updated covariance matrices of parameters for each layer \eqn{\Sigma_{\theta}}
}
}
\description{
This function feeds the neural network backward from responses to input data.
}
\seealso{
\code{\link{backwardHiddenStateUpdate}},
\code{\link{backwardParameterUpdate}}, \code{\link{forwardHiddenStateUpdate}}
}
