% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{feedForward}
\alias{feedForward}
\title{Forward uncertainty propagation}
\usage{
feedForward(NN, x, mp, Sp)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{x}{Input data}

\item{mp}{Mean vectors of parameters for each layer \eqn{\mu_{\theta}}}

\item{Sp}{Covariance matrices of parameters for each layer \eqn{\Sigma_{\theta}}}
}
\value{
\itemize{
\item Mean vectors of units for each layer \eqn{\mu_{Z}}
}

\itemize{
\item Covariance matrices of units for each layer \eqn{\Sigma_{Z}}
}

\itemize{
\item Covariance matrices between units and weights for each layer \eqn{\Sigma_{ZW}}
}

\itemize{
\item Covariance matrices between units and biases for each layer \eqn{\Sigma_{ZB}}
}

\itemize{
\item Covariance matrices between previous and current units for each layer \eqn{\Sigma_{ZZ^{+}}}
}
}
\description{
This function feeds the neural network forward from input data to
responses.
}
