% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{feedForwardPass}
\alias{feedForwardPass}
\title{Forward Uncertainty Propagation for Derivative Calculation}
\usage{
feedForwardPass(NN, theta, states)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}

\item{theta}{List of parameters}

\item{states}{List of states}
}
\value{
A list that contains:

\itemize{
\item \code{states}: List that contains states
}

\itemize{
\item \code{mda}: Mean vector of activation units' derivative
}

\itemize{
\item \code{Sda}: Covariance matrix of activation units' derivative
}

\itemize{
\item \code{mdda}: Mean vector of activation units' second derivative
}

\itemize{
\item \code{Sdda}: Covariance matrix of activation units' second derivative
}
}
\description{
This function feeds the neural network forward from input data to
responses and considers components required for derivative calculations.
}
