% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fowardHiddenStateUpdate}
\alias{fowardHiddenStateUpdate}
\title{Last hidden layer states update}
\usage{
fowardHiddenStateUpdate(mz, Sz, mzF, SzF, Cyz, y)
}
\arguments{
\item{mz}{Mean vector of the units for the output layer}

\item{Sz}{Covariance matrix of the units for the output layer}

\item{mzF}{Mean vector of the units for the output layer}

\item{SzF}{Covariance matrix of the units for the output layer}

\item{Cyz}{Covariance matrix between last hidden layer units and responses}

\item{y}{Response data}
}
\value{
A list that contains:

\itemize{
\item Updated mean vector of the last hidden layer units
}

\itemize{
\item Updated covariance matrix of the last hidden layer units
}
}
\description{
This function updates last hidden layer units using responses. It updates
\eqn{\mu_{z^{(0)}|y}} and \eqn{\Sigma_{z^{(0)}|y}} from the \eqn{z^{(0)}|y}
distribution.
}
