% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{globalParameterUpdate}
\alias{globalParameterUpdate}
\title{Backpropagation (Parameters Update)}
\usage{
globalParameterUpdate(theta, deltaTheta)
}
\arguments{
\item{theta}{List of parameters}

\item{deltaTheta}{Parameters' deltas (mean and covariance for each)}
}
\value{
List of updated parameters
}
\description{
This function updates parameters.
}
