% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{initialization}
\alias{initialization}
\title{Network initialization}
\usage{
initialization(NN)
}
\arguments{
\item{NN}{Lists the structure of the neural network}
}
\value{
\itemize{
\item NN with all required components
}

\itemize{
\item States of all required elements to perform TAGI
}
}
\description{
Verify and add components to the neural network structure.
}
