% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{initialization_net}
\alias{initialization_net}
\title{Network initialization}
\usage{
initialization_net(NN)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}
}
\value{
NN with all required components
}
\description{
Verify and add components to the neural network structure.
}
