% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{initializeInputs}
\alias{initializeInputs}
\title{Input initialization}
\usage{
initializeInputs(states, mz0, Sz0, ma0, Sa0, J0, mdxs0, Sdxs0, mxs0, Sxs0, xsc)
}
\arguments{
\item{states}{States of the neural network}

\item{mz0}{Input data}

\item{Sz0}{Variance of input data}

\item{ma0}{Activated input data}

\item{Sa0}{Variance of activated input data}

\item{J0}{Jacobian}

\item{...}{Other parameters}
}
\value{
States of the neural network
}
\description{
Initializes neural network inputs.
}
