% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{initializeWeightBias}
\alias{initializeWeightBias}
\title{Weights and biases initialization}
\usage{
initializeWeightBias(NN)
}
\arguments{
\item{NN}{Lists the structure of the neural network}
}
\value{
\itemize{
\item Initial mean vectors of parameters for each layer
}

\itemize{
\item Initial covariance matrices of parameters
for each layer
}
}
\description{
This function initializes the first weights and biases of the neural network.
}
