% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{initializeWeightBiasD}
\alias{initializeWeightBiasD}
\title{Weights and biases initialization for calculating derivatives}
\usage{
initializeWeightBiasD(NN)
}
\arguments{
\item{NN}{Lists the structure of the neural network}
}
\value{
All parameters required in the neural network to perform derivative calculations
}
\description{
This function initializes the first weights and biases of the neural network.
}
