% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{innovationVector}
\alias{innovationVector}
\title{Last hidden layer states' deltas update}
\usage{
innovationVector(SzF, dMz, dSz)
}
\arguments{
\item{SzF}{Covariance matrix of units for the next layer \eqn{\Sigma_{y}}}

\item{dMz}{Delta of mean vector of units for the next hidden layer \eqn{\mu_{Z}}}

\item{dSz}{Delta of covariance matrix of units for the next hidden layer \eqn{\Sigma_{Z}}}
}
\value{
\itemize{
\item Delta of mean vector of current hidden layer units given \eqn{y} \eqn{\mu_{Z}|y}
}

\itemize{
\item Delta of covariance matrix of current hidden layer units given \eqn{y} \eqn{\Sigma_{Z}|y}
}
}
\description{
This function updates hidden layer units' deltas using next hidden layer' deltas. It updates
\eqn{\mu_{Z|y}} and \eqn{\Sigma_{Z|y}} from the \eqn{Z|y}
distribution.
}
\details{
\eqn{f(\boldsymbol{z}|\boldsymbol{y}) = \mathcal{N}(\boldsymbol{z};\boldsymbol{\mu_{Z|y}},\boldsymbol{\Sigma_{Z^|y}})} where

\eqn{\boldsymbol{\mu_{{Z}|y}} = \boldsymbol{\mu_{Z}} + \boldsymbol{J_{Z}}(\boldsymbol{\mu_{Z^{+}|y}} - \boldsymbol{\mu_{Z^{+}}})}

\eqn{\boldsymbol{\Sigma_{{Z}|y}} = \boldsymbol{\Sigma_{Z}} + \boldsymbol{J_{Z}}(\boldsymbol{\Sigma_{Z^{+}|y}} - \boldsymbol{\Sigma_{Z^{+}}})\boldsymbol{J_{Z}^{T}}}

\eqn{\boldsymbol{J_{Z}} = \boldsymbol{\Sigma_{ZZ^{+}}} \boldsymbol{\Sigma_{Z^{+}}^{-1}}}
}
