% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{layerEncoder}
\alias{layerEncoder}
\title{Layer encoder}
\usage{
layerEncoder(NN)
}
\arguments{
\item{NN}{Lists the structure of the neural network}
}
\value{
NN with layer encoder
}
\description{
Add layer encoder to the neural network structure.
}
