% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{meanA}
\alias{meanA}
\title{Calculate mean of activated units}
\usage{
meanA(z, mz, funIdx)
}
\arguments{
\item{z}{Vector of units for the current layer}

\item{mz}{Mean vector of the units for the current layer}

\item{funIdx}{Activation function index defined by \code{\link{activationFunIndex}}}
}
\value{
A list that contains the activation units mean vector and the Jacobian
matrix evaluated at \eqn{\mu_{Z}}
}
\description{
This function uses lineratization to estimate the activation units mean vector
and the Jacobian matrix evaluated at mz.
}
