% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{meanMz}
\alias{meanMz}
\title{Mean vector of units}
\usage{
meanMz(mp, ma, idxFmwa, idxFmwab)
}
\arguments{
\item{mp}{Mean vector of parameters for the current layer}

\item{ma}{Mean vector of activation units from previous layer}

\item{idxFmwa}{Indices for weights and for activation
units for the current and previous layers respectively}

\item{idxFmwab}{Indices for biases of the current layer}
}
\value{
Mean vector of units for the current layer \eqn{\mu_{Z}}
}
\description{
This function calculate the mean vector of units \eqn{\mu_{Z}} for a given layer.
}
