% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{meanVarDev}
\alias{meanVarDev}
\title{Mean and Variance of Activated Units for Derivatives}
\usage{
meanVarDev(mz, Sz, funIdx, bound)
}
\arguments{
\item{mz}{Mean vector of the units for the current layer \eqn{\mu_{Z}}}

\item{Sz}{Covariance matrix of the units for the current layer \eqn{\Sigma_{Z}}}

\item{funIdx}{Activation function index defined by \code{\link{activationFunIndex}}}

\item{bound}{TBD}
}
\value{
The activation units' first derivative mean vector

The activation units' first derivative variance vector

The activation units' second derivative mean vector

The activation units' second derivative variance vector
}
\description{
This function calculates mean vector and covariance matrix of activation units'
derivatives.
}
