% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MedicalCost}
\alias{MedicalCost}
\title{Medical Cost of 1,338 insureds.}
\format{
A data frame with 1,338 rows and 10 variables:
\describe{
\item{age}{age of the insured}
\item{sex}{gender of the insured, binary (if female)}
\item{BMI}{Body Mass Index of the insured}
\item{children}{number of children covered as dependents}
\item{smoker}{smoking status, binary (if the insured smokes)}
\item{region: northeast}{binary (if the insured lives in that region)}
\item{region: southeast}{binary (if the insured lives in that region)}
\item{region: southwest}{binary (if the insured lives in that region)}
\item{region: northwest}{binary (if the insured lives in that region)}
\item{charges}{medical costs, in US dollars)}
}
}
\source{
\url{https://github.com/stedy/Machine-Learning-with-R-datasets/blob/master/insurance.csv}
}
\usage{
MedicalCost
}
\description{
A dataset containing the medical costs ("charges") and other attributes of
1,338 insureds.
}
\details{
The original dataset contains 7 variables, but one-hot encoding was
used on the "region" categorical variable.
}
\keyword{datasets}
