% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{network}
\alias{network}
\title{One iteration of the Tractable Approximate Gaussian Inference (TAGI)}
\usage{
network(NN, mp, Sp, x, y)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{mp}{Mean vector of parameters for each layer \eqn{\mu_{\theta}}}

\item{Sp}{Covariance matrix of parameters for each layer \eqn{\Sigma_{\theta}}}

\item{x}{Input data}

\item{y}{Response data}
}
\value{
\itemize{
\item Updated mean vector of parameters for each layer \eqn{\mu_{\theta}}
}

\itemize{
\item Updated covariance matrix of parameters for each layer \eqn{\Sigma_{\theta}}
}

\itemize{
\item Mean of predicted responses
}

\itemize{
\item Variance of the predicted responses
}
}
\description{
This function goes through one learning iteration of the neural network model
using TAGI.
}
