% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(xtrain, ytrain, xtest, ytest)
}
\arguments{
\item{xtrain}{Training set of input variables}

\item{ytrain}{Training set of responses}

\item{xtest}{Testing set of input variables}

\item{ytest}{Testing set of responses}
}
\value{
A list that contains:

\itemize{
\item The normalized training set of input variables
}

\itemize{
\item The normalized training set of responses
}

\itemize{
\item The normalized testing set of input variables
}

\itemize{
\item The normalized testing set of responses
}

\itemize{
\item Mean vector of input variables from training set
}

\itemize{
\item Variance vector of input variables from training set
}

\itemize{
\item Mean vector of responses from training set
}

\itemize{
\item Variance vector of responses from training set
}
}
\description{
This function normalizes data before entering the neural network.
}
