% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Regression problem}
\usage{
regression(NN, x, y, trainIdx, testIdx)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{x}{Input data}

\item{y}{Response data}

\item{trainIdx}{Observations IDs that are assigned to the training set}

\item{testIdx}{Observations IDs that are assigned to the testing set}
}
\value{
\itemize{
\item Mean vector of parameters for each layer \eqn{\mu_{\theta}}
}

\itemize{
\item Covariance matrix of parameters for each layer \eqn{\Sigma_{\theta}}
}

\itemize{
\item RMSE and LL metrics for each network models created
}

\itemize{
\item Training time of each neural network models created
}

\itemize{
\item Mean of predicted responses
}

\itemize{
\item Variance of the predicted responses
}
}
\description{
This function trains neural network models to solve a regression problem.
}
