% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{runBatchDerivative}
\alias{runBatchDerivative}
\title{Result of the TAGI with derivative calculations}
\usage{
runBatchDerivative(NN, xtrain, ytrain, xtest, ytest)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{xtrain}{Training set of input variables}

\item{ytrain}{Training set of responses}

\item{xtest}{Testing set of input variables}

\item{ytest}{Testing set of responses}
}
\value{
\itemize{
\item Mean of predicted responses
}

\itemize{
\item Variance of the predicted responses
}

\itemize{
\item Mean of first derivative of predicted responses
}

\itemize{
\item Mean of second derivative of predicted responses
}
}
\description{
This function returns the resulting derivatives from the neural network model using TAGI.
}
