% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{split}
\alias{split}
\title{Split data}
\usage{
split(x, y, ratio)
}
\arguments{
\item{x}{Input data}

\item{y}{Response data}

\item{ratio}{Training ratio}
}
\value{
A list that contains:

\itemize{
\item Training set of input variables
}

\itemize{
\item Training set of responses
}

\itemize{
\item Testing set of input variables
}

\itemize{
\item Testing set of responses
}
}
\description{
This function split data into training and test sets.
}
