% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToyExample.x_obs}
\alias{ToyExample.x_obs}
\title{Inputs used in training part for 1D toy problem}
\format{
A data frame with 20 rows and 1 variable \eqn{x}
}
\source{
\url{https://github.com/CivML-PolyMtl/TAGI/blob/master/ToyExample/ToyExample_1D.m}
}
\usage{
ToyExample.x_obs
}
\description{
The orignal dataset represents a 1D regression problem from Hernández-Lobato & Adams (2015):
\eqn{y = x^{3} + \epsilon} where \eqn{\epsilon \sim \mathcal{N}(0,9)} and \eqn{x \in [\,-4,4]\,}. In this dataset, \eqn{x} and \eqn{y} are normalized.
}
\details{
The dataset generated with the seed from the TAGI repository was used for comparison purposes.
}
\references{
Hernández-Lobato, J. M., & Adams, R. "Probabilistic
backpropagation for scalable learning of bayesian neural networks."
International Conference on Machine Learning. 2015.
}
\keyword{datasets}
