% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{activationFunIndex}
\alias{activationFunIndex}
\title{Assign ID to activation functions}
\usage{
activationFunIndex(funName)
}
\arguments{
\item{funName}{Type of activation function: "tanh", "sigm", "cdf", "relu" or
"softplus"}
}
\value{
An ID number which corresponds to:

\itemize{
\item 1 if \code{funName} is "tanh"
}

\itemize{
\item 2 if \code{funName} is "sigm"
}

\itemize{
\item 3 if \code{funName} is "cdf"
}

\itemize{
\item 4 if \code{funName} is "relu"
}

\itemize{
\item 5 if \code{funName} is "softplus"
}
}
\description{
This function assigns an ID number depending on the type of activation function.
}
