% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{backwardHiddenStateUpdate}
\alias{backwardHiddenStateUpdate}
\title{Backward hidden states update}
\usage{
backwardHiddenStateUpdate(mz, Sz, mzF, SzF, SzB, Czz, mzB, idx)
}
\arguments{
\item{mz}{Mean vector of the units for the current layer}

\item{Sz}{Covariance matrix of the units for the current layer}

\item{mzF}{Mean vector of the units for the next layer}

\item{SzF}{Covariance matrix of the units for the next layer}

\item{SzB}{Updated covariance matrix of the units for the next layer}

\item{Czz}{Covariance matrix between units of the previous and currents layers}

\item{mzB}{Updated mean vector of the units for the next layer}

\item{idx}{List that contains the indices for the hidden state update step of
the current layer}
}
\value{
A list that contains:

\itemize{
\item Updated mean vector of the units for the current layer
}

\itemize{
\item Updated covariance matrix of the units for the current layer
}
}
\description{
This function updates hidden units from responses to input data. It updates
\eqn{\mu_{z|y}} and \eqn{\Sigma_{z|y}} from the \eqn{z|y}
distribution for a given layer.
}
