% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{backwardParameterUpdate}
\alias{backwardParameterUpdate}
\title{Backward parameters update}
\usage{
backwardParameterUpdate(mp, Sp, mzF, SzF, SzB, Czp, mzB, idx)
}
\arguments{
\item{mp}{Mean vector of parameters for the current layer \eqn{\mu_{\theta}}}

\item{Sp}{Covariance matrix of parameters for the current layer \eqn{\Sigma_{\theta}}}

\item{mzF}{Mean vector of units for the next layer \eqn{\mu_{Z^{+}}}}

\item{SzF}{Covariance matrix of units for the next layer \eqn{\Sigma_{Z^{+}}}}

\item{SzB}{Covariance matrix of units for the next layer given \eqn{y} \eqn{\Sigma_{Z^{+}|y}}}

\item{Czp}{Covariance matrix between units and parameters for the current layer \eqn{\Sigma_{\theta Z^{+}}}}

\item{mzB}{Mean vector of units for the next layer given \eqn{y} \eqn{\mu_{Z^{+}|y}}}

\item{idx}{Indices for the parameter update step of
the current layer}
}
\value{
\itemize{
\item Mean vector of parameters for the current layer given \eqn{y} \eqn{\mu_{\theta|y}}
}

\itemize{
\item Covariance matrix of parameters for the current layer given \eqn{y} \eqn{\Sigma_{\theta|y}}
}
}
\description{
This function updates parameters from responses to input data. It updates
\eqn{\mu_{\theta|y}} and \eqn{\Sigma_{\theta|y}} from the \eqn{\theta|y}
distribution for a given layer.
}
\details{
\eqn{f(\boldsymbol{\theta}|\boldsymbol{y}) = \mathcal{N}(\boldsymbol{\theta};\boldsymbol{\mu_{\theta|y}},\boldsymbol{\Sigma_{\theta|y}})} where

\eqn{\boldsymbol{\mu_{\theta|y}} =\boldsymbol{\mu_{\theta}} + \boldsymbol{J_{\theta}}(\boldsymbol{\mu_{Z^{+}|y}} - \boldsymbol{\mu_{Z^{+}}})}

\eqn{\boldsymbol{\Sigma_{\theta|y}} = \boldsymbol{\Sigma_{\theta}} + \boldsymbol{J_{\theta}}(\boldsymbol{\Sigma_{Z^{+}|y}} - \boldsymbol{\Sigma_{Z^{+}}})\boldsymbol{J_{\theta}^{T}}}

\eqn{\boldsymbol{J_{\theta}} = \boldsymbol{\Sigma_{\theta Z^{+}}}\boldsymbol{\Sigma^{-1}_{Z^{+}}}}
}
