% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{batchDerivative}
\alias{batchDerivative}
\title{One iteration of the Tractable Approximate Gaussian Inference (TAGI) with Derivative Calculations}
\usage{
batchDerivative(NN, theta, normStat, states, x, Sx, y, dlayer)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}

\item{theta}{List of parameters}

\item{normStat}{TBD}

\item{states}{List of states}

\item{x}{Set of input data}

\item{Sx}{Variance of x}

\item{y}{Set of corresponding responses}

\item{dlayer}{Layer from which derivatives will be in respect to}
}
\value{
A list that contains:

\itemize{
\item \code{theta}: List of updated parameters
}

\itemize{
\item \code{normStat}: updated TBD
}

\itemize{
\item \code{mzl}: Predicted responses
}

\itemize{
\item \code{Szl}: Variance vector of predicted responses
}

\itemize{
\item \code{mdg}: Predicted derivatives
}

\itemize{
\item \code{Sdg}: Variance vector of derivatives
}

\itemize{
\item \code{Cdgz}: Covariance between derivatives and inputs
}

\itemize{
\item \code{mddg}: Predicted second derivatives
}
}
\description{
This function goes through one learning iteration of the neural network model
using TAGI.
}
