% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{catParameters}
\alias{catParameters}
\title{Concatenate Parameters}
\usage{
catParameters(mw, Sw, mb, Sb, mwx, Swx, mbx, Sbx)
}
\arguments{
\item{mw}{TBD}

\item{Sw}{TBD}

\item{mb}{TBD}

\item{Sb}{TBD}

\item{mwx}{TBD}

\item{Swx}{TBD}

\item{mbx}{TBD}

\item{Sbx}{TBD}
}
\value{
mw TBD in vector format

Sw TBD in vector format

mb TBD in vector format

Sb TBD in vector format

mwx TBD in vector format

Swx TBD in vector format

mbx TBD in vector format

Sbx TBD in vector format
}
\description{
Combines in a single column vector each parameter for all layers.
}
