% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{compressParameters}
\alias{compressParameters}
\title{Compress Parameters}
\usage{
compressParameters(mw, Sw, mb, Sb, mwx, Swx, mbx, Sbx)
}
\arguments{
\item{mw}{TBD}

\item{Sw}{TBD}

\item{mb}{TBD}

\item{Sb}{TBD}

\item{mwx}{TBD}

\item{Swx}{TBD}

\item{mbx}{TBD}

\item{Sbx}{TBD}
}
\value{
theta List of parameters
}
\description{
Put together parameters into a list of parameters.
}
