% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{compressStates}
\alias{compressStates}
\title{Compress States}
\usage{
compressStates(mz, Sz, ma, Sa, J, mdxs, Sdxs, mxs, Sxs)
}
\arguments{
\item{mz}{TBD}

\item{Sz}{TBD}

\item{ma}{TBD}

\item{Sa}{TBD}

\item{J}{TBD}

\item{mdxs}{TBD}

\item{Sdxs}{TBD}

\item{mxs}{TBD}

\item{Sxs}{TBD}
}
\value{
states List of states
}
\description{
Put together states into a list of states.
}
