% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt.R
\name{computeError}
\alias{computeError}
\title{Compute Error}
\usage{
computeError(y, ypred)
}
\arguments{
\item{y}{A vector or a matrix of the real y's}

\item{ypred}{A vector or a matrix of the predicted y's form the model}
}
\value{
A number which is the RMSE
}
\description{
This function calculates the Root Mean Square Error (RMSE). It takes as input
two vectors (or matrices) with one containing the real y's and the other the
predicted y's from the model.
}
