% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{covariance}
\alias{covariance}
\title{Indices for covariances in the neural network}
\usage{
covariance(NN)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}
}
\value{
NN with new elements:

\itemize{
\item idxFmwa: Lists of indices (weights and activation units) for
deterministic matrix F *  \eqn{\mu_{WA}} for each layer
}

\itemize{
\item idxFmwab: List of bias indices for deterministic matrix F *  \eqn{\mu_{B}} for
each layer
}

\itemize{
\item idxFCzwa: Lists of indices (weights and activation units) for
deterministic matrix F *  \eqn{\Sigma_{ZWA}} for each layer
}

\itemize{
\item idxSzpUd: List of indices for the parameter update step for each layer
}

\itemize{
\item idxSzzUd: List of indices for the hidden state update step for each
layer
}

\itemize{
\item idxFCwwa: Lists of indices (weights and activation units) for deterministic matrix F *  \eqn{\Sigma_{WA\theta}} for
each layer
}

\itemize{
\item idxa: List of indices for activation unit for each layer
}

\itemize{
\item idxFCb: List of bias indices for deterministic matrix F *  \eqn{\Sigma_{C}} for
each layer
}
}
\description{
This function assigns indices for all covariance elements in
the neural network.
}
