% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{covarianceCzp}
\alias{covarianceCzp}
\title{Covariance matrices between units and parameters}
\usage{
covarianceCzp(ma, Sp, idxFCwwa, idxFCb)
}
\arguments{
\item{ma}{Mean vector of the activation units from previous layer \eqn{\mu_{A}}}

\item{Sp}{Covariance matrix of the parameters for the current layer \eqn{\Sigma_{\theta}}}

\item{idxFCwwa}{List that contains the indices for weights and for activation
units for the current and previous layers respectively}

\item{idxFCb}{Indices for biases of the current layer}
}
\value{
A list that contains:

\itemize{
\item Covariance matrix between units and biases for the current layer \eqn{\Sigma_{ZB}}
}

\itemize{
\item Covariance matrix between units and weights for the current layer \eqn{\Sigma_{ZW}}
}
}
\description{
This function calculate the covariance matrices between units and parameters
\eqn{\Sigma_{ZW}} and \eqn{\Sigma_{ZB}} for a given layer.
}
