% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{covarianceCzz}
\alias{covarianceCzz}
\title{Covariance matrix between units of the previous and current layers}
\usage{
covarianceCzz(mp, Sz, J, idxCawa)
}
\arguments{
\item{mp}{Mean vector of parameters for the current layer \eqn{\mu_{\theta}}}

\item{Sz}{Covariance matrix of units for the current layer \eqn{\Sigma_{Z}}}

\item{J}{Jacobian matrix evaluated at \eqn{\mu_{Z}}}

\item{idxCawa}{Indices for weights and for activation
units for the current and previous layers respectively}
}
\value{
Covariance matrix between units of previous and current layers \eqn{\Sigma_{ZZ^{+}}}
}
\description{
This function calculate the covariance matrix between units of the previous
and current layers \eqn{\Sigma_{ZZ^{+}}} for a given layer.
}
