% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{covarianceSa}
\alias{covarianceSa}
\title{Calculate variance of activated units}
\usage{
covarianceSa(J, Sz)
}
\arguments{
\item{J}{Jacobian matrix evaluated at \eqn{\mu_{Z}}}

\item{Sz}{Covariance matrix of units for the current layer \eqn{\Sigma_{Z}}}
}
\value{
The activation units covariance matrix \eqn{\Sigma_{A}}
}
\description{
This function uses lineratization to estimate the covariance matrix of activation units \eqn{\Sigma_{A}}.
}
