% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{covarianceSa}
\alias{covarianceSa}
\title{Calculate variance of activated units}
\usage{
covarianceSa(J, Sz)
}
\arguments{
\item{J}{Jacobian matrix evaluated at \eqn{\mu_{z}}}

\item{Sz}{Covariance matrix of the units for the current layer}
}
\value{
The activation units variance vector
}
\description{
This function uses lineratization to estimate the activation units variance.
}
