% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{covdx}
\alias{covdx}
\title{Covariance between Derivatives and Inputs}
\usage{
covdx(mwo, mw, mdgo2, mpdi, mdgoe, Cdozi, Cdizi, ni, no, no2, B)
}
\arguments{
\item{mwo}{Mean vector of the weights for the next layer}

\item{mw}{Mean vector of the weights for the current layer}

\item{mdgo2}{Mean vector of derivatives in 2nd next layer}

\item{mpdi}{TBD}

\item{mdgoe}{Unit vector of length  (# response variables * B)}

\item{Cdozi}{Covariance between derivative (next layer) and hidden (current layer) layers}

\item{Cdizi}{between derivative and hidden layers (same layer)}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}
}
\value{
Covariance between derivative and input layers
}
\description{
This function calculates covariance between derivatives and inputs.
}
