% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{derivative}
\alias{derivative}
\title{Derivative Calculation}
\usage{
derivative(NN, theta, states, mda, Sda, dlayer)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}

\item{theta}{List that of parameters}

\item{states}{List of states}

\item{mda}{Mean vector of activation units' derivative}

\item{Sda}{Covariance matrix of activation units' derivative}

\item{dlayer}{Layer from which derivatives will be in respect to}
}
\value{
A list that contains:

\itemize{
\item \code{mdg}: Mean vector of derivative of the output layer
}

\itemize{
\item \code{Sdg}: Variance of derivative of the output layer
}

\itemize{
\item \code{Cdgz}: Covariance of derivative of the output layer
}
}
\description{
This function does derivative calculations.
}
