% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{extractParameters}
\alias{extractParameters}
\title{Extract parameters}
\usage{
extractParameters(theta)
}
\arguments{
\item{theta}{List of parameters}
}
\value{
\itemize{
\item Mean vector of weights for the current layer
}

\itemize{
\item Covariance vector of weights for the current layer
}

\itemize{
\item Mean vector of biases for the current layer
}

\itemize{
\item Covariance vector of biases for the current layer
}
}
\description{
Extract parameters from list of parameters.
}
