% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{extractStates}
\alias{extractStates}
\title{Extract States}
\usage{
extractStates(mz, Sz, ma, Sa, J, mdxs, Sdxs, mxs, Sxs)
}
\arguments{
\item{states}{List of states}
}
\value{
mz TBD

Sz TBD

ma TBD

Sa TBD

J TBD

mdxs TBD

Sdxs TBD

mxs TBD

Sxs TBD
}
\description{
Extract states from list of states.
}
