% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCombinaisonDweight}
\alias{fcCombinaisonDweight}
\title{Combination of products of first derivative (iterations on weights)}
\usage{
fcCombinaisonDweight(mpdi, mw, Sw, mda, Sda, ni, no, B)
}
\arguments{
\item{mpdi}{Mean vector of first derivative product wd of current layer}

\item{mw}{Mean vector of weights for the current layer}

\item{Sw}{Covariance of weights for the current layer}

\item{mda}{Mean vector of activation units' first derivative from current layer}

\item{Sda}{Covariance of the activation units' first derivative from current layer}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Mean array of combination of products of first derivatives
}
\description{
This function calculates mean of combination of products of first derivatives (wd)*(wd).
Each weight is multiplied to another weight (including itself) from the same node.
Each node is multiplied to the same node (in the same layer).
}
