% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCombinaisonDweightNodeAll}
\alias{fcCombinaisonDweightNodeAll}
\title{All Possible Combinations of Products of First Order Derivatives}
\usage{
fcCombinaisonDweightNodeAll(mpdi, mpdin, mpdiw, ni, no, B)
}
\arguments{
\item{mpdi}{Mean matrix of the current product of derivative and weight}

\item{mpdin}{Mean array of combination of products of first order derivatives (iterations on nodes)}

\item{mpdiw}{Mean array of combination of products of first order derivatives (iterations on weights)}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Mean array of combination of products of first order derivatives
}
\description{
This function calculates mean of products of first order derivatives wd*wd.
Since both weight and node are iterated over all products, every products
(weight times node) from current layer are considered which results in a
(Bni x no x noni)-array. I.e. each dimension of the array represents a single
product being multiplied to all other possible products from current layer.
Order is as followed: w11d1, w12d2, w13d3, ..., w1nidni, w21d1, w22d2, ..., w2nidni, ... wno1d1, ..., wnonidni
}
