% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovawaa}
\alias{fcCovawaa}
\title{Covariance between activation units and weights}
\usage{
fcCovawaa(mw, Sw, Jo, mai, Sai, ni, no, B)
}
\arguments{
\item{mw}{Mean vector of weights for the current layer}

\item{Sw}{Covariance of weights for the current layer}

\item{Jo}{Jacobian of next layer}

\item{mai}{Mean vector of activation units from current layer}

\item{Sai}{Covariance of activation units from current layer}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
\itemize{
\item Covariance between activation units and weights
}

\itemize{
\item Covariance between activation units from current and next layers
}
}
\description{
This function calculates covariance between activation units and weights and
covariance between activation units from consecutive layers.
}
