% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovaz}
\alias{fcCovaz}
\title{Covariance between Activation and Hidden Units}
\usage{
fcCovaz(mw, Sw, mda, Sda, ni, no, B)
}
\arguments{
\item{mw}{Mean vector of the weights for the current layer}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}

\item{Jo}{Jacobian of next layer}

\item{J}{Jacobian of current layer}

\item{Sz}{Covariance of the units from current layer}
}
\value{
Covariance between activation and hidden layers (same layer)

Covariance between activation (next layer) and hidden (current layer) layers
}
\description{
This function calculates covariance between activation and hidden units.
}
