% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovdaddd}
\alias{fcCovdaddd}
\title{Covariance between First and Second Orders Derivatives}
\usage{
fcCovdaddd(mao, mai, mdai, Caoai, Cdodi, acti, ni, no, B)
}
\arguments{
\item{mao}{Mean vector of the activation units from next layer}

\item{mai}{Mean vector of the activation units from current layer}

\item{mdai}{Mean vector of the activation units' derivative from current layer}

\item{Caoai}{Covariance between activation units from current and next layers}

\item{Cdodi}{Covariance between derivatives from current and next layers}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Covariance between first order derivatives from next layer and activation units from current layer

Covariance between first and second orders derivatives from consecutive layers
}
\description{
This function calculates covariance between first order derivatives and activation units and
covariance between first and second orders derivatives from consecutive layers.
}
