% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovdz}
\alias{fcCovdz}
\title{Covariance between Derivatives and Hidden Units}
\usage{
fcCovdz(mao, mai, Caizi, Caozi, acto, acti, ni, no, B)
}
\arguments{
\item{mao}{Mean vector of the activation units from next layer}

\item{mai}{Mean vector of the activation units from current layer}

\item{Caizi}{Covariance between activation and hidden layers (same layer)}

\item{Caozi}{Covariance between activation (next layer) and hidden (current layer) layers}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Covariance between derivative (next layer) and hidden (current layer) layers

Covariance between derivative and hidden layers (same layer)
}
\description{
This function calculates covariance between derivatives and hidden units.
}
