% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovwdo2wdiwdi}
\alias{fcCovwdo2wdiwdi}
\title{Covariance between products in (same) next and current layers}
\usage{
fcCovwdo2wdiwdi(mpdo, Cwdowdiwdi)
}
\arguments{
\item{mpdo}{Mean vector of first derivative product wd of next layer}

\item{Cwdowdiwdi}{Covariance cov(wdo,wdi*wdi) of weights times derivatives products terms when there are one product in next layer and two in current}
}
\value{
Covariance cov(wdo^2,wdi*wdi) of weights times derivatives products terms when there are two products in both next and current layers
}
\description{
This function calculates covariance cov(wdo^2,wdiwdi) of weights times derivatives
products terms when there are two products in both next and current layers.
The product fom next layer is the same squared.
}
