% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovwdowdi2}
\alias{fcCovwdowdi2}
\title{Covariance between Next Layer Product and Current Layer Squared Product}
\usage{
fcCovwdowdi2(mpdi, Cdgoddgik)
}
\arguments{
\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{Cdgoddgik}{Covariance between weights times derivatives from consecutive layers}
}
\value{
Covariance cov(wdo,(wdi)^2) of weights times derivatives products terms when there is one product in next layer and two squared in current
}
\description{
This function calculates covariance cov(wdo,(wdi)^2) of weights times derivatives
products terms when there are one product in next layer and two squared in current.
}
