% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCovwdowdiwdi}
\alias{fcCovwdowdiwdi}
\title{Covariance between Next Layer Product and Current Layer Multiplied Products}
\usage{
fcCovwdowdiwdi(mpdi, Cdgoddgik, ni, no, B)
}
\arguments{
\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{Cdgoddgik}{Covariance between weights times derivatives from consecutive layers}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Covariance cov(wdo,wdi*wdi) of weights times derivatives products terms when there is one product in next layer and two in current
}
\description{
This function calculates covariance cov(wdo,wdi*wdi) of weights times derivatives
products terms when there are one product in next layer and two in current.
}
