% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcCwdowdowwdi2}
\alias{fcCwdowdowwdi2}
\title{Covariance between Next Layer Multiplied Products and Current Layer Multiplied Products (Same Derivative)}
\usage{
fcCwdowdowwdi2(mpdi, mpdo, Cdgodgi, acti, ni, no, no2, B)
}
\arguments{
\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{mpdo}{Mean vector of the first order derivative product wd of next layer}

\item{Cdgodgi}{Covariance between weights times derivatives from consecutive layers}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}
}
\value{
Covariance cov(wdowdo,wdiwdi) where the di terms are the same
}
\description{
This function calculates covariance cov(wdowdo,wdiwdi) where the di terms are the same, when next second layer involves only a product term (wddo2).
}
