% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDerivative}
\alias{fcDerivative}
\title{Derivatives for Fully Connected layers}
\usage{
fcDerivative(
  mw,
  Sw,
  mwo,
  Jo,
  J,
  mao,
  Sao,
  mai,
  Sai,
  Szi,
  mdai,
  Sdai,
  mdgo,
  mdgoe,
  Sdgo,
  mdgo2,
  acto,
  acti,
  ni,
  no,
  no2,
  B
)
}
\arguments{
\item{mw}{Mean vector of the weights for the current layer}

\item{Sw}{Covariance of the weights for the current layer}

\item{mwo}{Mean vector of the weights for the next layer}

\item{Jo}{Jacobian of next layer}

\item{J}{Jacobian of current layer}

\item{mao}{Mean vector of the activation units from next layer}

\item{Sao}{Covariance of the activation units from next layer}

\item{mai}{Mean vector of the activation units from current layer}

\item{Sai}{Covariance of the activation units from current layer}

\item{Szi}{Covariance of the units from current layer}

\item{mdai}{Mean vector of the activation units' derivative from current layer}

\item{Sdai}{Covariance of the activation units' derivative from current layer}

\item{mdgo}{Mean vector of derivatives in next layer}

\item{mdgoe}{Mean derivatives at each node in next layer}

\item{Sdgo}{Variance of derivatives in next layer}

\item{mdgo2}{Mean vector of derivatives in 2nd next layer}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}
}
\value{
Mean vector of the derivatives

Covariance matrix of the derivatives
}
\description{
This function calculates mean and variance of derivatives and covariance of derivative and input layers.
}
