% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDerivative2}
\alias{fcDerivative2}
\title{Second derivatives for fully connected layers}
\usage{
fcDerivative2(
  mw,
  mwo,
  mao,
  mai,
  mdai,
  mddai,
  mpddi,
  mdgo,
  mdgo2,
  Caoai,
  Cdow,
  Cdodi,
  acti,
  ni,
  no,
  no2,
  B
)
}
\arguments{
\item{mw}{Mean vector of weights for the current layer}

\item{mwo}{Mean vector of weights for the next layer}

\item{mao}{Mean vector of activation units from next layer}

\item{mai}{Mean vector of activation units from current layer}

\item{mdai}{Mean vector of activation units' first derivative from current layer}

\item{mddai}{Mean vector of activation units' second derivative from current layer}

\item{mpddi}{Mean vector of second derivative product wdd of current layer}

\item{mdgo}{Mean vector of product of derivatives in next layer}

\item{mdgo2}{Mean vector of product of derivatives in second next layer}

\item{Caoai}{Covariance between activation units from current and next layers}

\item{Cdow}{Covariance between first derivatives and weights}

\item{Cdodi}{Covariance between first derivatives from current and next layers}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}
}
\value{
Mean of product terms for second derivative calculations
}
\description{
This function calculates mean of product of derivatives, when new product term involves second derivatives (wdd).
}
