% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcDerivative3}
\alias{fcDerivative3}
\title{Products of first derivatives multiplied to second derivative for fully connected layers}
\usage{
fcDerivative3(
  mw,
  Sw,
  mwo,
  mao,
  mai,
  mdao,
  mdai,
  Sdai,
  mpdi,
  mdgo,
  mdgo2,
  Caow,
  Caoai,
  Cdow,
  Cdodi,
  acto,
  acti,
  ni,
  no,
  no2,
  B,
  dlayer
)
}
\arguments{
\item{mw}{Mean vector of weights for the current layer}

\item{Sw}{Covariance of weights for the current layer}

\item{mwo}{Mean vector of weights for the next layer}

\item{mao}{Mean vector of activation units from next layer}

\item{mai}{Mean vector of activation units from current layer}

\item{mdao}{Mean vector of activation units' first derivative from next layer}

\item{mdai}{Mean vector of activation units' first derivative from current layer}

\item{Sdai}{Covariance of activation units' first derivative from current layer}

\item{mpdi}{Mean vector of first derivative product wd of current layer}

\item{mdgo}{Mean vector of product of derivatives in next layer}

\item{mdgo2}{Mean vector of product of derivatives in second next layer}

\item{Caow}{Covariance between activation units and weights}

\item{Caoai}{Covariance between activation units from current and next layers}

\item{Cdow}{Covariance between first derivatives and weights}

\item{Cdodi}{Covariance between first derivatives from current and next layers}

\item{acto}{Activation function index for next layer defined by \code{\link{activationFunIndex}}}

\item{acti}{Activation function index for current layer defined by \code{\link{activationFunIndex}}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in second next layer}

\item{B}{Batch size}

\item{dlayer}{TRUE if derivatives will be in respect to current layer}
}
\value{
Mean of product terms for second derivative calculations
}
\description{
This function calculates mean of product of derivatives, when new product term involves product
of two first derivatives (wdwd) from the same layer multiplied to second derivatives (wdd) from next layer.
}
