% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcHiddenStateBackwardPass}
\alias{fcHiddenStateBackwardPass}
\title{Backpropagation (Deltas) for Fully Connected layers (Many Observations)}
\usage{
fcHiddenStateBackwardPass(Sz, Sxs, J, mw, deltaM, deltaS, ni, no, B, rB)
}
\arguments{
\item{Sz}{Covariance of the units from current layer}

\item{Sxs}{TBD}

\item{J}{Jacobian of current layer}

\item{mw}{Mean vector of the weights for the current layer}

\item{deltaM}{Delta of mean vector of the next hidden layer units given \eqn{y} \eqn{\mu_{Z}|y}}

\item{deltaS}{Delta of covariance matrix of the next hidden layer units given \eqn{y} \eqn{\Sigma_{Z}|y}}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}

\item{rB}{Number of times batch size is repeated}
}
\value{
\itemize{
\item Delta of mean vector of the hidden layer units given \eqn{y} \eqn{\mu_{Z}|y}}
}

\itemize{
\item Delta of covariance matrix of the hidden layer units given \eqn{y} \eqn{\Sigma_{Z}|y}}
}

\itemize{
\item TBD
}

\itemize{
\item TBD
}
}
\description{
This function update last hidden layer's deltas.
}
