% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcMeanDlayer2}
\alias{fcMeanDlayer2}
\title{Mean of Weights times Derivatives Products Terms Squared}
\usage{
fcMeanDlayer2(mpdi, mpdi2, mdgo, Cdgoddgik, ni, no, no2, B)
}
\arguments{
\item{mpdi}{Mean vector of the first order derivative product wd of current layer}

\item{mdgo}{Mean vector of derivatives in next layermx Mean vector of inputs}

\item{Cdgoddgik}{Covariance between weights times derivatives from consecutive layers}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{no2}{Number of units in 2nd next layer}

\item{B}{Batch size}

\item{mpddi}{Mean vector of the second order derivative product wdd of current layer}
}
\value{
Mean of weights times derivatives products terms squared
}
\description{
This function calculates mean and variance of weights times derivatives products terms squared.
}
