% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{fcMeanVarDnode}
\alias{fcMeanVarDnode}
\title{Mean and Covariance Vectors of Derivatives}
\usage{
fcMeanVarDnode(mw, Sw, mda, Sda, ni, no, B)
}
\arguments{
\item{mw}{Mean vector of the weights for the current layer}

\item{Sw}{Covariance of the weights for the current layer}

\item{mda}{Mean vector of the activation units' derivative from current layer}

\item{Sda}{Covariance of the activation units' derivative from current layer}

\item{ni}{Number of units in current layer}

\item{no}{Number of units in next layer}

\item{B}{Batch size}
}
\value{
Mean vector of the derivatives

Covariance matrix of the derivatives
}
\description{
This function calculates the mean vector and the covariance matrix for derivatives.
}
