% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{feedForward}
\alias{feedForward}
\title{Forward uncertainty propagation}
\usage{
feedForward(NN, x, mp, Sp)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}

\item{x}{Set of input data}

\item{mp}{List that contains mean vectors of the parameters for each layer}

\item{Sp}{List that contains covariance matrices of the
parameters for each layer}
}
\value{
A list that contains:

\itemize{
\item mz: List that contains the mean vectors of the units
for each layer
}

\itemize{
\item Sz: List that contains the covariance matrices of the
units for each layer
}

\itemize{
\item Czw: List that contains the covariance matrices between
units and weights for each layer
}

\itemize{
\item Czb: List that contains the covariance matrices between
units and biases for each layer
}

\itemize{
\item Czz: List that contains the covariance matrices between
previous and current units for each layer
}
}
\description{
This function feeds the neural network forward from input data to
responses.
}
