% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{feedForward}
\alias{feedForward}
\title{Forward Uncertainty Propagation}
\usage{
feedForward(NN, x, mp, Sp)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}

\item{x}{Set of input data}

\item{mp}{List that contains mean vectors of the parameters for each layer \eqn{\mu_{\theta}}}

\item{Sp}{List that contains covariance matrices of the
parameters for each layer \eqn{\Sigma_{\theta}}}
}
\value{
A list that contains:

\itemize{
\item \code{mz}: List that contains the mean vectors of the units
for each layer \eqn{\mu_{Z}}
}

\itemize{
\item \code{Sz}: List that contains the covariance matrices of the
units for each layer \eqn{\Sigma_{Z}}
}

\itemize{
\item \code{Czw}: List that contains the covariance matrices between
units and weights for each layer \eqn{\Sigma_{ZW}}
}

\itemize{
\item \code{Czb}: List that contains the covariance matrices between
units and biases for each layer \eqn{\Sigma_{ZB}}
}

\itemize{
\item \code{Czz}: List that contains the covariance matrices between
previous and current units for each layer \eqn{\Sigma_{ZZ^{+}}}
}
}
\description{
This function feeds the neural network forward from input data to
responses.
}
