% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{feedForwardPass}
\alias{feedForwardPass}
\title{Forward uncertainty propagation for derivative calculation}
\usage{
feedForwardPass(NN, theta, states)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{theta}{List of parameters}

\item{states}{List of states}
}
\value{
\itemize{
\item Updated states
}

\itemize{
\item Mean vectors of activation units' first derivative
}

\itemize{
\item Covariance matrices of activation units' first derivative
}

\itemize{
\item Mean vectors of activation units' second derivative
}

\itemize{
\item Covariance matrices of activation units' second derivative
}
}
\description{
This function feeds the neural network forward from input data to
responses and considers components required for derivative calculations.
}
