% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{forwardHiddenStateUpdate}
\alias{forwardHiddenStateUpdate}
\title{Last hidden layer states update}
\usage{
forwardHiddenStateUpdate(mz, Sz, mzF, SzF, Cyz, y)
}
\arguments{
\item{mz}{Mean vector of units for the last hidden layer \eqn{\mu_{X^{(0)}}}}

\item{Sz}{Covariance matrix of units for the last hidden layer \eqn{\Sigma_{Z^{(0)}}}}

\item{mzF}{Mean vector of units for the output layer \eqn{\mu_{y}}}

\item{SzF}{Covariance matrix of tunits for the output layer \eqn{\Sigma_{y}}}

\item{Cyz}{Covariance matrix between last hidden layer units and responses \eqn{\Sigma_{YZ^{(0)}}}}

\item{y}{Response data}
}
\value{
\itemize{
\item Mean vector of last hidden layer units given \eqn{y} \eqn{\mu_{Z^{(0)}|y}}
}

\itemize{
\item Covariance matrix of last hidden layer units given \eqn{y} \eqn{\Sigma_{Z^{(0)}|y}}
}
}
\description{
This function updates last hidden layer units using responses. It updates
\eqn{\mu_{Z^{(0)}|y}} and \eqn{\Sigma_{Z^{(0)}|y}} from the \eqn{Z^{(0)}|y}
distribution.
}
\details{
\eqn{f(\boldsymbol{z}^{(0)}|\boldsymbol{y}) = \mathcal{N}(\boldsymbol{z}^{(0)};\boldsymbol{\mu_{Z^{(0)}|y}},\boldsymbol{\Sigma_{Z^{(0)}|y}})} where

\eqn{\boldsymbol{\mu_{{Z}^{(0)}|y}} = \boldsymbol{\mu_{Z^{(0)}}} + \boldsymbol{\Sigma_{YZ^{(0)}}^{T}}\boldsymbol{\Sigma^{-1}_{Y}}(\boldsymbol{y} - \boldsymbol{\mu_{Y}})}

\eqn{\boldsymbol{\Sigma_{{Z}^{(0)}|y}} = \boldsymbol{\Sigma_{Z^{(0)}}} - \boldsymbol{\Sigma_{YZ^{(0)}}^{T}}\boldsymbol{\Sigma^{-1}_{Y}}\boldsymbol{\Sigma_{YZ^{(0)}}}}
}
