% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{hiddenStateBackwardPass}
\alias{hiddenStateBackwardPass}
\title{Backpropagation (States' Deltas)}
\usage{
hiddenStateBackwardPass(NN, theta, states, y, Sy, udIdx)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}

\item{theta}{List that of parameters}

\item{states}{List of states}

\item{y}{A vector or a matrix of responses}

\item{Sy}{Variance of responses}

\item{udIdx}{TBD}
}
\value{
A list that contains:

\itemize{
\item Delta of mean vector of units given \eqn{y} \eqn{\mu_{Z}|y}} at all layers
}

\itemize{
\item Delta of covariance matrix of units given \eqn{y} \eqn{\Sigma_{Z}|y}} at all layers
}
}
\description{
This function calculates states' deltas.
}
