% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{initialization}
\alias{initialization}
\title{Network initialization}
\usage{
initialization(NN)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}
}
\value{
NN NN with all required components

states States of all required elements to perform TAGI
}
\description{
Verify and add components to the neural network structure.
}
