% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{initializeStates}
\alias{initializeStates}
\title{States Initialization}
\usage{
initializeStates(nodes, B, rB, xsc)
}
\arguments{
\item{nodes}{Vector that contains the number of nodes at each layer}

\item{B}{Batch size}

\item{rB}{Number of times batch size is repeated}

\item{xsc}{TBD}
}
\value{
states: States of the neural network
}
\description{
Initiliazes neural network states.
}
