% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{initializeWeightBias}
\alias{initializeWeightBias}
\title{Weights and biases initialization}
\usage{
initializeWeightBias(NN)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}
}
\value{
A list that contains:

\itemize{
\item List that contains initial mean vectors of the parameters for each layer
}

\itemize{
\item List that contains initial covariance matrices of the parameters
for each layer
}
}
\description{
This function initializes the first weights and biases of the neural network.
}
