% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt.R
\name{loglik}
\alias{loglik}
\title{Compute log-likelihood}
\usage{
loglik(y, ypred, Vpred)
}
\arguments{
\item{y}{Response data}

\item{ypred}{Mean of predicted responses}

\item{Vpred}{Variance of the predicted responses}
}
\value{
LL for the given data
}
\description{
This function calculates the log-likelihood (LL). It takes as input
three vectors (or matrices) with one containing the real \eqn{y}'s, one with the
predicted \eqn{y}'s from the model and the last one with the variance of the \eqn{y}'s.
}
