% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{meanVar}
\alias{meanVar}
\title{Mean, Jacobian and variance of activated units}
\usage{
meanVar(z, mz, Sz, funIdx)
}
\arguments{
\item{z}{Vector of units for the current layer}

\item{mz}{Mean vector of units for the current layer \eqn{\mu_{Z}}}

\item{Sz}{Covariance matrix of units for the current layer \eqn{\Sigma_{Z}}}

\item{funIdx}{Activation function index defined by \code{\link{activationFunIndex}}}
}
\value{
\itemize{
\item Mean vector of activation units for the current layer \eqn{\mu_{A}}
}

\itemize{
\item Covariance matrix activation units for the current layer \eqn{\Sigma_{A}}
}

\itemize{
\item Jacobian matrix evaluated at \eqn{\mu_{Z}}
}
}
\description{
This function returns mean vector \eqn{\mu_{A}}, Jacobian matrix evaluated at
\eqn{\mu_{Z}} and covariance matrix of activation units \eqn{\Sigma_{A}}.
}
