% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{meanVarDev}
\alias{meanVarDev}
\title{Mean and variance of activated units for derivatives}
\usage{
meanVarDev(mz, Sz, funIdx, bound)
}
\arguments{
\item{mz}{Mean vector of units for the current layer \eqn{\mu_{Z}}}

\item{Sz}{Covariance matrix of units for the current layer \eqn{\Sigma_{Z}}}

\item{funIdx}{Activation function index defined by \code{\link{activationFunIndex}}}

\item{bound}{If layer is bound}
}
\value{
\itemize{
\item Mean vector of activation units' first derivative
}

\itemize{
\item Covariance matrix of activation units' first derivative
}

\itemize{
\item Mean vector activation units' second derivative
}

\itemize{
\item Covariance matrix activation units' second derivative
}
}
\description{
This function calculates mean vector and covariance matrix of activation units'
derivatives.
}
