% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagi.R
\name{network}
\alias{network}
\title{One iteration of the Tractable Approximate Gaussian Inference (TAGI)}
\usage{
network(NN, mp, Sp, x, y)
}
\arguments{
\item{NN}{List that contains the structure of the neural network}

\item{mp}{List that contains mean vectors of the parameters for each layer}

\item{Sp}{List that contains covariance matrices of the
parameters for each layer}

\item{x}{Set of input data}

\item{y}{Set of corresponding responses}
}
\value{
A list that contains:

\itemize{
\item mp: List that contains the updated mean vectors of the parameters
for each layer
}

\itemize{
\item Sp: List that contains the updated covariance matrices of the
parameters for each layer
}

\itemize{
\item zn: Predicted responses
}

\itemize{
\item Szn: Variance vector of predicted responses
}
}
\description{
This function goes through one learning iteration of the neural network model
using TAGI.
}
