% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Regression problem}
\usage{
regression(NN, x, y, trainIdx, testIdx)
}
\arguments{
\item{NN}{Lists the structure of the neural network}

\item{x}{Set of input data}

\item{y}{Set of corresponding responses}

\item{trainIdx}{Observations IDs that are assigned to the training set}

\item{testIdx}{Observations IDs that are assigned to the testing set}
}
\value{
A list that contains:

\itemize{
\item \code{mp}: List that contains the updated mean vectors of the parameters
for each layer \eqn{\mu_{\theta}}
}

\itemize{
\item \code{Sp}: List that contains the updated covariance matrices of the
parameters for each layer \eqn{\Sigma_{\theta}}
}

\itemize{
\item \code{metric}: List that contains RMSE and LL metrics for each neural
network models created
}

\itemize{
\item \code{time}: Training time of each neural network models created
}
}
\description{
This function trains neural network models to solve a regression problem.
}
